+++ pipelines/maintenance__db_retention/tasks.py
@@ -6,35 +6,38 @@ from prefect.client.schemas.filters import FlowRunFilter, FlowRunFilterState, Fl
 from prefect.client.schemas.objects import StateType
 from prefect.exceptions import ObjectNotFound
 
-
 @task
-async def delete_stale_pending_runs(
-    threshold_hours: int = 1,
+async def delete_old_flow_runs(
+    days_to_keep: int = 25,
     batch_size: int = 200
 ):
     """Delete completed flow runs older than specified days."""
     # logger = get_run_logger()
 
     async with get_client() as client:
-        cutoff = datetime.now(timezone.utc) - timedelta(hours=threshold_hours)
+        cutoff = datetime.now(timezone.utc) - timedelta(days=days_to_keep)
 
-        # Create filter for pending runs older than threshold
+        # Create filter for old completed flow runs
         # Note: Using start_time because created time filtering is not available
         flow_run_filter = FlowRunFilter(
             start_time=FlowRunFilterStartTime(before_=cutoff),
             state=FlowRunFilterState(
                 type=FlowRunFilterStateType(
-                    any_=[StateType.PENDING]
+                    any_=[
+                        StateType.COMPLETED,
+                        StateType.FAILED,
+                        StateType.CANCELLED,
+                        StateType.CRASHED,
+                    ]
                 )
-            )
+            ),
         )
 
         # Get flow runs to delete
-        flow_runs = await client.read_flow_runs(
-            flow_run_filter=flow_run_filter,
-            limit=batch_size
-        )
-
+        flow_runs = await client.read_flow_runs(flow_run_filter=flow_run_filter, limit=batch_size)
+        to_delete_dates = [run.start_time if run.start_time else run.created for run in flow_runs]
+        to_delete_dates.sort()
+        print(to_delete_dates)
         deleted_total = 0
 
         while flow_runs:
@@ -47,13 +50,12 @@ async def delete_stale_pending_runs(
                     await client.delete_flow_run(flow_run.id)
                     deleted_total += 1
                     batch_deleted += 1
-                    print(f"Deleted pending flow run: {flow_run.name}")
                 except ObjectNotFound:
                     # Already deleted (e.g., by concurrent cleanup) - treat as success
                     deleted_total += 1
                     batch_deleted += 1
                 except Exception as e:
-                    print(f"Failed to delete flow run {flow_run.name}: {e}")
+                    print(f"Failed to delete flow run {flow_run.id}: {e}")
                     failed_deletes.append(flow_run.id)
 
                 # # Rate limiting - adjust based on your API capacity
@@ -66,11 +68,11 @@ async def delete_stale_pending_runs(
 
             # Get next batch
             flow_runs = await client.read_flow_runs(
-                flow_run_filter=flow_run_filter,
-                limit=batch_size
+                flow_run_filter=flow_run_filter, limit=batch_size
             )
 
             # Delay between batches to avoid overwhelming the API
             await asyncio.sleep(0.5)
 
-        print(f"Janitoring complete. Total deleted: {deleted_total}")
+        print(f"Retention complete. Total deleted: {deleted_total}")
+
